% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{associate_device_with_placement}
\alias{associate_device_with_placement}
\title{Associates a physical device with a placement}
\usage{
associate_device_with_placement(projectName, placementName, deviceId,
  deviceTemplateName)
}
\arguments{
\item{projectName}{[required] The name of the project containing the placement in which to associate the device.}

\item{placementName}{[required] The name of the placement in which to associate the device.}

\item{deviceId}{[required] The ID of the physical device to be associated with the given placement in the project. Note that a mandatory 4 character prefix is required for all \code{deviceId} values.}

\item{deviceTemplateName}{[required] The device template name to associate with the device ID.}
}
\description{
Associates a physical device with a placement.
}
\section{Accepted Parameters}{
\preformatted{associate_device_with_placement(
  projectName = "string",
  placementName = "string",
  deviceId = "string",
  deviceTemplateName = "string"
)
}
}

