% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{create_placement}
\alias{create_placement}
\title{Creates an empty placement}
\usage{
create_placement(placementName, projectName, attributes = NULL)
}
\arguments{
\item{placementName}{The name of the placement to be created.}

\item{projectName}{The name of the project in which to create the placement.}

\item{attributes}{Optional user-defined key/value pairs providing contextual data (such as location or function) for the placement.}
}
\description{
Creates an empty placement.
}
\examples{

}
