% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{delete_placement}
\alias{delete_placement}
\title{Deletes a placement}
\usage{
delete_placement(
  placementName = "string",
  projectName = "string"
)
}
\arguments{
\item{placementName}{[required] The name of the empty placement to delete.}

\item{projectName}{[required] The project containing the empty placement to delete.}
}
\description{
Deletes a placement. To delete a placement, it must not have any devices associated with it.
}
\details{
When you delete a placement, all associated data becomes irretrievable.
}
\examples{

}
