% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{delete_project}
\alias{delete_project}
\title{Deletes a project}
\usage{
delete_project(
  projectName = "string"
)
}
\arguments{
\item{projectName}{[required] The name of the empty project to delete.}
}
\description{
Deletes a project. To delete a project, it must not have any placements associated with it.
}
\details{
When you delete a project, all associated data becomes irretrievable.
}
\examples{

}
