% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{describe_placement}
\alias{describe_placement}
\title{Describes a placement in a project}
\usage{
describe_placement(placementName, projectName)
}
\arguments{
\item{placementName}{The name of the placement within a project.}

\item{projectName}{The project containing the placement to be described.}
}
\description{
Describes a placement in a project.
}
\examples{

}
