% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{describe_project}
\alias{describe_project}
\title{Returns an object describing a project}
\usage{
describe_project(projectName)
}
\arguments{
\item{projectName}{[required] The name of the project to be described.}
}
\description{
Returns an object describing a project.
}
\section{Accepted Parameters}{
\preformatted{describe_project(
  projectName = "string"
)
}
}

