% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{disassociate_device_from_placement}
\alias{disassociate_device_from_placement}
\title{Removes a physical device from a placement}
\usage{
disassociate_device_from_placement(projectName, placementName,
  deviceTemplateName)
}
\arguments{
\item{projectName}{The name of the project that contains the placement.}

\item{placementName}{The name of the placement that the device should be removed from.}

\item{deviceTemplateName}{The device ID that should be removed from the placement.}
}
\description{
Removes a physical device from a placement.
}
\examples{

}
