% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{get_devices_in_placement}
\alias{get_devices_in_placement}
\title{Returns an object enumerating the devices in a placement}
\usage{
get_devices_in_placement(projectName, placementName)
}
\arguments{
\item{projectName}{[required] The name of the project containing the placement.}

\item{placementName}{[required] The name of the placement to get the devices from.}
}
\description{
Returns an object enumerating the devices in a placement.
}
\section{Accepted Parameters}{
\preformatted{get_devices_in_placement(
  projectName = "string",
  placementName = "string"
)
}
}

