% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{list_placements}
\alias{list_placements}
\title{Lists the placement(s) of a project}
\usage{
list_placements(projectName, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{projectName}{The project containing the placements to be listed.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return per request. If not set, a default value of 100 is used.}
}
\description{
Lists the placement(s) of a project.
}
\examples{

}
