% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{list_projects}
\alias{list_projects}
\title{Lists the AWS IoT 1-Click project(s) associated with your AWS account and region}
\usage{
list_projects(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return per request. If not set, a default value of 100 is used.}
}
\description{
Lists the AWS IoT 1-Click project(s) associated with your AWS account and region.
}
\section{Accepted Parameters}{
\preformatted{list_projects(
  nextToken = "string",
  maxResults = 123
)
}
}

