% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_package.R
\docType{package}
\name{paws.iot1clickprojects-package}
\alias{paws.iot1clickprojects}
\alias{paws.iot1clickprojects-package}
\title{paws.iot1clickprojects: AWS IoT 1-Click Projects Service}
\description{
The AWS IoT 1-Click Project API Reference
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
