% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iot1clickprojects_operations.R
\name{update_placement}
\alias{update_placement}
\title{Updates a placement with the given attributes}
\usage{
update_placement(
  placementName = "string",
  projectName = "string",
  attributes = list(
    "string"
  )
)
}
\arguments{
\item{placementName}{[required] The name of the placement to update.}

\item{projectName}{[required] The name of the project containing the placement to be updated.}

\item{attributes}{The user-defined object of attributes used to update the placement. The maximum number of key/value pairs is 50.}
}
\description{
Updates a placement with the given attributes. To clear an attribute, pass an empty value (i.e., "").
}
\examples{

}
